requirements.txt
# Make sure to run with python3.10 (3.11 gives errors with sqlite)
# Since it is a large repo, there are a lot of dependencies. 
# You might not need all of them, but just in case we list them all.
git+https://github.com/LLaVA-VL/LLaVA-NeXT.git
torch>=2.5.1
torchvision>=0.20.1
accelerate>=0.26.0
torchaudio>=2.5.1
torchcodec
transformers>=4.47
tokenizers>=0.20.0
info-nce-pytorch>=0.1.1
opencv-python>=4.7.0
decord>=0.6.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
pydantic>=2.0.0
tqdm>=4.65.0
ffmpeg-python>=0.2.0
pathlib>=1.0.1
requests>=2.31.0
vllm>0.7.2
open-clip-torch
av<13 # Gave errors with av 13
InstructorEmbedding
sentence-transformers>3.4.0
huggingface-hub[hf_transfer]
soundfile
qwen-vl-utils
sentencepiece
bitsandbytes>=0.45.2
moviepy
matplotlib
submitit
conette
resampy
pytorchvideo
webdataset
jsonlines
openai
python-dotenv
einops
polars
tenacity
wandb
loguru
aac-metrics
peft
pytorch-lightning
trl>=0.16.0
datasets
xgrammar
cuda-python
nvidia-cuda-nvrtc-cu12
ninja
pybind11
fire
deepspeed
wget
h5py
einops_exts
blobfile